requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
anthropic==0.5.0
google-api-python-client==2.97.0
python-dotenv==1.0.0